fix: address review feedback on test categorization
- Remove broad except Exception in test_real_web_connection
- Move test_real_local_connection from external/ to live/
