+++ tests/external/test_server_connection.py
@@ -1,4 +1,4 @@
-"""Integration tests for real Zotero connections (if credentials available)."""
+"""Integration test for real Zotero web API connection."""
 
 import os
 
@@ -29,26 +29,5 @@ class TestZoteroConnectionIntegration:
         assert client.mode == "web"
         assert isinstance(client._client, zotero.Zotero)
 
-        try:
-            items = client._client.items(limit=1)
-            assert isinstance(items, list)
-        except Exception as e:
-            pytest.fail(f"Failed to connect to real Zotero API: {e}")
-
-    def test_real_local_connection(self) -> None:
-        """Test connection to local Zotero server (if available)."""
-        if os.getenv("TEST_ZOTERO_LOCAL", "false").lower() != "true":
-            pytest.skip("Local Zotero not configured")
-
-        local_settings = Settings(
-            zotero_local=True,
-            zotero_library_id="",
-            zotero_api_key=None,
-        )
-
-        client = ZoteroClient(local_settings)
-
-        assert client.mode == "local"
-        assert isinstance(client._client, zotero.Zotero)
-        assert client._client.local is True
-        assert client._client.endpoint == "http://localhost:23119/api"
+        items = client._client.items(limit=1)
+        assert isinstance(items, list)
